delphes data properly processed
