+++ scripts/local_test_delphes.sh
@@ -25,12 +25,12 @@ cd test_tmp_delphes
 python3 ../mlpf/pytorch/graph_data_delphes.py --dataset data/delphes_cfi \
   --processed_dir data/delphes_cfi/processed --num-files-merge 1 --num-proc 1
 
-# #run the pytorch training
-# COMET_API_KEY="bla" python3 ../mlpf/pytorch/train_end2end_farouk.py \
-#   --dataset data/TTbar_14TeV_TuneCUETP8M1_cfi --space_dim 2 --n_train 3 \
-#   --n_val 2 --model PFNet7 --convlayer gravnet-radius --convlayer2 sgconv \
-#   --lr 0.0001 --hidden_dim 32 --n_epochs 2 --l1 1.0 --l2 0.001 --target cand \
-#   --batch_size 1 --dropout 0.2 --disable_comet
+#run the pytorch training
+COMET_API_KEY="bla" python3 ../mlpf/pytorch/train_end2end_delphes.py \
+  --dataset data/delphes_cfi --space_dim 2 --n_train 2 \
+  --n_val 1 --model PFNet7 --convlayer gravnet-radius --convlayer2 sgconv \
+  --lr 0.0001 --hidden_dim 32 --n_epochs 2 --l1 1.0 --l2 0.001 --target cand \
+  --batch_size 1 --dropout 0.2 --disable_comet
 #
 # #generate dataframe with predictions from the pytorch model
 # python3 ../mlpf/pytorch/eval_end2end.py --dataset data/TTbar_14TeV_TuneCUETP8M1_cfi \
